Moves TracedModule to tf_test_utils.py
